+++ lib/locatine/public.rb
@@ -1,0 +1,103 @@
+module Locatine
+  ##
+  # Public methods of the Search class
+  module Public
+    ##
+    # Creates a new instance of Search
+    #
+    # Params:
+    # +json+ is the name of file to store//read data. Default =>
+    # "./Locatine_files/default.json"
+    #
+    # +depth+ is the value that shows how many data will be stored for element.
+    #
+    # +browser+ is the instance of Watir::Browser. Unless provided it gonna
+    # be created with locatine-app onboard.
+    #
+    # +learn+ shows will locatine ask for assistance from user or will fail
+    # on error. learn is true when LEARN parameter is set in environment.
+    #
+    # +stability_limit+ shows max times attribute should be present to
+    # consider it trusted.
+    #
+    # +scope+ will be used in search (if not provided) defaulkt is "Default"
+    def initialize(json: './Locatine_files/default.json',
+                   depth: 3,
+                   browser: nil,
+                   learn: ENV['LEARN'].nil? ? false : true,
+                   stability_limit: 10,
+                   scope: 'Default')
+      browser ||= right_browser
+      @browser = browser
+      @json = json
+      @folder = File.dirname(@json)
+      @name = File.basename(@json)
+      @depth = depth
+      @data = read_create
+      @learn = learn
+      @stability_limit = stability_limit
+      @scope = scope
+    end
+
+    ##
+    # Looking for the element
+    #
+    # Params:
+    #
+    # +scope+ is a parameter that is used to get information about the
+    # element from @data. Default is "Default"
+    #
+    # +name+ is a parameter that is used to get information about the
+    # element from @data. Must not be nil.
+    #
+    # +exact+ if true locatine will be forced to use only basic search.
+    # Default is false
+    #
+    # +locator+ if not empty it is used for the first attempt to find the
+    # element. Default is {}
+    #
+    # +vars+ hash of variables that will be used for dynamic attributes.
+    # See readme for example
+    #
+    # +look_in+ only elements of that kind will be used. Use Watir::Browser
+    # methods returning collections (:text_fields, :links, :divs, etc.)
+    #
+    # +iframe+ if provided locatine will look for elements inside of it
+    #
+    # +return_locator+ is to return a valid locator of the result
+    #
+    # +collection+ when true an array will be returned. When false - a
+    # single element
+    def find(simple_name = nil,
+             name: nil,
+             scope: nil,
+             exact: false,
+             locator: {},
+             vars: {},
+             look_in: nil,
+             iframe: nil,
+             return_locator: false,
+             collection: false)
+      name = set_name(simple_name, name)
+      @type = look_in
+      @iframe = iframe
+      scope ||= @scope.nil? ? 'Default' : @scope
+      result, attributes = full_search(name, scope, vars, locator, exact)
+      return { xpath: generate_xpath(attributes, vars) } if result &&
+                                                            return_locator
+      return to_subtype(result, collection) if result && !return_locator
+    end
+
+    ##
+    # Find alias with return_locator option enforced
+    def lctr(*args)
+      enforce(:return_locator, true, *args)
+    end
+
+    ##
+    # Find alias with collection option enforced
+    def collect(*args)
+      enforce(:collection, true, *args)
+    end
+  end
+end